+++ src/lanscape/ui/blueprints/__init__.py
@@ -1,4 +1,4 @@
-from ..libraries.subnet_scan import ScanManager
+from ...libraries.subnet_scan import ScanManager
 import logging
 # defining here so blueprints can access the same
 # manager instance
