+++ src/lanscape/ui/blueprints/api/tools.py
@@ -1,8 +1,8 @@
 from flask import request, jsonify
 from . import api_bp
-from ...libraries.net_tools import get_all_network_subnets
-from ...libraries.ip_parser import parse_ip_input
-from ...libraries.errors import SubnetTooLargeError
+from ....libraries.net_tools import get_all_network_subnets
+from ....libraries.ip_parser import parse_ip_input
+from ....libraries.errors import SubnetTooLargeError
 import traceback
